Move generate random payload attestation
