+++ testing/util/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "deposits.go",
         "helpers.go",
         "merge.go",
+        "payload_attestation.go",
         "state.go",
         "sync_aggregate.go",
         "sync_committee.go",
